% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcplr.R
\name{tcplr}
\alias{tcplr}
\title{Tutorials of the class taught by Prof.Lu Jie using R}
\usage{
tcplr(classid)
}
\arguments{
\item{classid}{A Character with double quotation marks}
}
\value{
A RShiny output file
}
\description{
Here are tutorials for Prof.Lu Jie's class "Political Data Analysis and Programming", written by the Teaching Assistant Liu Tianxiang.
}
